ref: refactor listing entity
The listing entity was refactored. Validation annotations were removed and added
to ListingDTO. Also, createListing method was refactored to accept ListingDTO as
the parameter and Listing as the returned value instead of the previous implementation
that made it the other way around.
