+++ src/main/java/com/codeplanks/home360/event/listener/ListingEnquiryEventListener.java
@@ -1,10 +1,13 @@
+/* (C)2024 */
 package com.codeplanks.home360.event.listener;
 
-import com.codeplanks.home360.service.EmailServiceImpl;
-import com.codeplanks.home360.event.ListingEnquiryEvent;
-import com.codeplanks.home360.service.ListingService;
 import com.codeplanks.home360.domain.listing.ListingWithAgentInfo;
+import com.codeplanks.home360.event.ListingEnquiryEvent;
+import com.codeplanks.home360.service.EmailServiceImpl;
+import com.codeplanks.home360.service.ListingServiceImpl;
 import jakarta.mail.MessagingException;
+import java.io.UnsupportedEncodingException;
+import java.util.Locale;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
@@ -14,17 +17,13 @@ import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Component;
 import org.thymeleaf.context.Context;
 
-import java.io.UnsupportedEncodingException;
-import java.util.Locale;
-
 @Component
 @Slf4j
 @RequiredArgsConstructor
 public class ListingEnquiryEventListener implements ApplicationListener<ListingEnquiryEvent> {
   private final JavaMailSender mailSender;
-  private final ListingService listingService;
+  private final ListingServiceImpl listingService;
   private final EmailServiceImpl emailService;
-
 
   @Value("${application.frontend.user-listings-url.url}")
   private String userListingUrl;
@@ -44,15 +43,16 @@ public class ListingEnquiryEventListener implements ApplicationListener<ListingE
     }
   }
 
-  public void sendEnquiryNotificationEmail(String listingUrl, String agentEmail,
-                                           String agentFirstName) throws MessagingException,
-          UnsupportedEncodingException {
+  public void sendEnquiryNotificationEmail(
+      String listingUrl, String agentEmail, String agentFirstName)
+      throws MessagingException, UnsupportedEncodingException {
     Context context = new Context(Locale.ENGLISH);
     String subject = "Listing Enquiry Notification";
 
-    String mailContent = "You have a new enquiry about one of your listings.Login to your " +
-            "dashboard to view the content.The listing being enquired about can be found in the " +
-            "url below.";
+    String mailContent =
+        "You have a new enquiry about one of your listings.Login to your "
+            + "dashboard to view the content.The listing being enquired about can be found in the "
+            + "url below.";
 
     context.setVariable("name", agentFirstName);
     context.setVariable("subject", subject);
